Refactor core to centralize SDS and separate questionnaire responses
