+++ src/app/core/lib/questionnaire-response/index.ts
@@ -1,0 +1,1 @@
+export * from './questionnaire-response';
