+++ src/app/core/mapping/category-mappings.json
@@ -12,8 +12,8 @@
         }
         ,
          {
-            "name": "questionaires",
-            "value": "questionaires",
+            "name": "questionnaires",
+            "value": "questionnaires",
             "type": "category"
         }  
         ,
