+++ src/app/health-status/health-status.component.html
@@ -24,6 +24,6 @@
     <app-clinical-therapy-results *ngIf="featureToggling.status.therapy"></app-clinical-therapy-results>
   </section>
   <section class="clearfix">
-    <app-clinical-questionaires-results *ngIf="featureToggling.status.questionaires"></app-clinical-questionaires-results>
+    <app-clinical-questionnaires-results *ngIf="featureToggling.status.questionnaires"></app-clinical-questionnaires-results>
   </section>
 </div>
